CHG,ADD: MergeDatasets.merge() calls the new MergeModels.merge(),
also the MatchingGraphUtils inner class was moved as-is to MergeModels
